pyproject.toml
[project]
name = "staykin-be"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "greenlet>=3.5.0",
    "imagekitio>=5.4.0",
    "pydantic-settings>=2.14.0",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
